music_web/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.1.5
gunicorn==20.1.0
django-environ==0.10.0
whitenoise==6.4.0
psycopg2-binary==2.9.6
Pillow==9.5.0
python-vlc